upgrade create task

+++ src/main/java/pl/kanban/kanban_table/service/KanbanSerImp.java
@@ -4,8 +4,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import pl.kanban.kanban_table.model.Task;
 import pl.kanban.kanban_table.repository.TaskRepo;
-
 import java.util.List;
+import java.util.Optional;
 
 @Controller
 public class KanbanSerImp implements KanbanService {
@@ -21,12 +21,20 @@ public class KanbanSerImp implements KanbanService {
     }
 
     @Override
-    public void addTask(Task task) {
-        taskRepo.save(task);
+    public void deleteTask(long id) {
+        Optional<Task> deleteTask = taskRepo.findById(id);
+        if(deleteTask.isPresent()) taskRepo.deleteById(id);
     }
 
     @Override
-    public void deleteTask(long id) {
-        taskRepo.deleteById(id);
+    public void addTask(Task task) {
+        Optional<Task> oldTask = taskRepo.findById(task.getId());
+        if(oldTask.isPresent()){
+            Task updateTask=oldTask.get();
+            updateTask.setTaskTitle(task.getTaskTitle());
+            updateTask.setTaskTable(task.getTaskTable());
+            taskRepo.save(updateTask);
+        }
+        else taskRepo.save(task);
     }
 }
